Moved block library under insight
