+++ src/insight/block.js
